fix: decouple rich text editor & decorator
