+++ packages/draft-editor/src/block-renderer-fn.type.ts
@@ -1,13 +1,10 @@
-import { ContentBlock, ContentState, DraftDecoratorType } from 'draft-js'
-import { RichTextEditorComponent } from './draft-editor.type'
+import { ContentBlock, ContentState } from 'draft-js'
 
 export type AtomicBlockProps<T> = {
   block: ContentBlock
   blockProps: {
     onEditStart: () => void
     onEditFinish: (arg0?: { entityKey?: string; entityData?: T }) => void
-    RichTextEditorComponent: RichTextEditorComponent
-    decorator: DraftDecoratorType
   }
   contentState: ContentState
 }
