+++ packages/draft-editor/src/buttons/info-box.tsx
@@ -2,15 +2,18 @@ import React, { useState } from 'react'
 import buttonNames from './bt-names'
 import {
   AtomicBlockUtils,
-  DraftDecoratorType,
   EditorState,
   RawDraftContentState,
   convertToRaw,
   convertFromRaw,
 } from 'draft-js'
 import { Drawer, DrawerController } from '@keystone-ui/modals'
-import { RichTextEditorProps } from '../draft-editor.type'
+import { RichTextEditor } from '../rich-text-editor'
 import { Select } from './form/select'
+import { editableAnnotationDecorator } from '../entity-decorators/annotation'
+import { editableLinkDecorator } from '../entity-decorators/link'
+
+const editableDecorators = [editableAnnotationDecorator, editableLinkDecorator]
 
 const disabledButtons = [
   buttonNames.h2,
@@ -49,18 +52,16 @@ type InfoBoxInputType = {
     rawContentState: RawDraftContentState
   }) => void
   onCancel: () => void
-  Editor: React.ComponentType<RichTextEditorProps>
-  decorator: DraftDecoratorType
   inputValue: InfoBoxInputValue
 }
 
 export function InfoBoxInput(props: InfoBoxInputType) {
-  const { isOpen, onConfirm, onCancel, Editor, decorator, inputValue } = props
+  const { isOpen, onConfirm, onCancel, inputValue } = props
 
   const contentState = convertFromRaw(inputValue.rawContentState)
   const [inputValueState, setInputValueState] = useState({
     type: inputValue.type,
-    editorState: EditorState.createWithContent(contentState, decorator),
+    editorState: EditorState.createWithContent(contentState),
   })
 
   return (
@@ -111,7 +112,8 @@ export function InfoBoxInput(props: InfoBoxInputType) {
             })
           }}
         />
-        <Editor
+        <RichTextEditor
+          decorators={editableDecorators}
           disabledButtons={disabledButtons}
           editorState={inputValueState.editorState}
           onChange={(editorState: EditorState) => {
@@ -132,75 +134,65 @@ type InfoBoxButtonProps = {
   onChange: (param: EditorState) => void
 }
 
-export function createInfoBoxButton({
-  InnerEditor,
-  decorator,
-}: {
-  InnerEditor: React.ComponentType<RichTextEditorProps>
-  decorator: DraftDecoratorType
-}) {
-  return function InfoBoxButton(props: InfoBoxButtonProps) {
-    const [toShowInput, setToShowInput] = useState(false)
-    const { className, editorState, onChange: onEditorStateChange } = props
+export const InfoBoxButton = (props: InfoBoxButtonProps) => {
+  const [toShowInput, setToShowInput] = useState(false)
+  const { className, editorState, onChange: onEditorStateChange } = props
 
-    const onChange = ({
-      type,
-      rawContentState,
-    }: {
-      type: InfoBoxTypeEnum
-      rawContentState: RawDraftContentState
-    }) => {
-      const contentState = editorState.getCurrentContent()
+  const onChange = ({
+    type,
+    rawContentState,
+  }: {
+    type: InfoBoxTypeEnum
+    rawContentState: RawDraftContentState
+  }) => {
+    const contentState = editorState.getCurrentContent()
 
-      // create an InfoBox entity
-      const contentStateWithEntity = contentState.createEntity(
-        'INFOBOX',
-        'IMMUTABLE',
-        {
-          type,
-          rawContentState,
-        }
-      )
-      const entityKey = contentStateWithEntity.getLastCreatedEntityKey()
-      const newEditorState = EditorState.set(editorState, {
-        currentContent: contentStateWithEntity,
-      })
-
-      //The third parameter here is a space string, not an empty string
-      //If you set an empty string, you will get an error: Unknown DraftEntity key: null
-      onEditorStateChange(
-        AtomicBlockUtils.insertAtomicBlock(newEditorState, entityKey, ' ')
-      )
-      setToShowInput(false)
-    }
-
-    return (
-      <React.Fragment>
-        {toShowInput && (
-          <InfoBoxInput
-            Editor={InnerEditor}
-            decorator={decorator}
-            onConfirm={onChange}
-            onCancel={() => {
-              setToShowInput(false)
-            }}
-            isOpen={toShowInput}
-            inputValue={{
-              type: InfoBoxTypeEnum.newsChargeStation,
-              rawContentState: { blocks: [], entityMap: {} },
-            }}
-          />
-        )}
-        <div
-          className={className}
-          onClick={() => {
-            setToShowInput(true)
-          }}
-        >
-          <i className="far"></i>
-          <span>InfoBox</span>
-        </div>
-      </React.Fragment>
+    // create an InfoBox entity
+    const contentStateWithEntity = contentState.createEntity(
+      'INFOBOX',
+      'IMMUTABLE',
+      {
+        type,
+        rawContentState,
+      }
     )
+    const entityKey = contentStateWithEntity.getLastCreatedEntityKey()
+    const newEditorState = EditorState.set(editorState, {
+      currentContent: contentStateWithEntity,
+    })
+
+    //The third parameter here is a space string, not an empty string
+    //If you set an empty string, you will get an error: Unknown DraftEntity key: null
+    onEditorStateChange(
+      AtomicBlockUtils.insertAtomicBlock(newEditorState, entityKey, ' ')
+    )
+    setToShowInput(false)
   }
+
+  return (
+    <React.Fragment>
+      {toShowInput && (
+        <InfoBoxInput
+          onConfirm={onChange}
+          onCancel={() => {
+            setToShowInput(false)
+          }}
+          isOpen={toShowInput}
+          inputValue={{
+            type: InfoBoxTypeEnum.newsChargeStation,
+            rawContentState: { blocks: [], entityMap: {} },
+          }}
+        />
+      )}
+      <div
+        className={className}
+        onClick={() => {
+          setToShowInput(true)
+        }}
+      >
+        <i className="far"></i>
+        <span>InfoBox</span>
+      </div>
+    </React.Fragment>
+  )
 }
