--- packages/draft-editor/src/draft-editor.type.ts
@@ -1,10 +1,0 @@
-import React from 'react'
-import { EditorState } from 'draft-js'
-
-export type RichTextEditorProps = {
-  onChange: (editorState: EditorState) => void
-  editorState: EditorState
-  disabledButtons?: string[]
-}
-
-export type RichTextEditorComponent = React.ComponentType<RichTextEditorProps>
